isletix_langraph/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.7",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
